// types/clients.ts
export interface Client {
    code: string;
    name: string;
}

export interface ClientData {
    cli_des: string;   
    co_cli: string;    
    co_zon?: string;    
    dir_ent2?: string;  
    direc1?: string;    
    direc2?: string;  
    rif: string;
    telefonos: string;
    email: string;

}