// features/createOrder/interfaces/pedidoDTO.ts
import { RenglonPedidoDTO } from "./ReglonPedidoDTO";
export interface PedidoDTO {
    fact_num: number;
    contrib?: boolean;
    nombre?: string;
    rif?: string;
    nit?: string;
    status?: string;
    comentario?: string;
    descrip?: string;
    saldo?: number;
    fec_emis?: string;  // ISO date
    fec_venc?: string;
    co_cli?: string;
    co_ven?: string;
    co_tran?: string;
    dir_ent?: string;
    forma_pag?: string;
    tot_bruto?: number;
    tot_neto?: number;
    glob_desc?: number;
    tot_reca?: number;
    porc_gdesc?: string;
    porc_reca?: string;
    total_uc?: number;
    total_cp?: number;
    tot_flete?: number;
    monto_dev?: number;
    totklu?: number;
    anulada?: boolean;
    impresa?: boolean;
    iva?: number;
    iva_dev?: number;
    feccom?: string;
    numcom?: number;
    tasa?: number;
    moneda?: string;
    cta_contab?: string;
    seriales?: number;
    tasag?: number;
    tasag10?: number;
    tasag20?: number;
    campo1?: string;
    campo2?: string;
    campo3?: string;
    campo4?: string;
    campo5?: string;
    campo6?: string;
    campo7?: string;
    campo8?: string;
    co_us_in?: string;
    fe_us_in?: string;
    co_us_mo?: string;
    fe_us_mo?: string;
    co_us_el?: string;
    fe_us_el?: string;
    revisado?: string;
    trasnfe?: string;
    co_sucu?: string;
    rowguid?: string;
    mon_ilc?: number;
    otros1?: number;
    otros2?: number;
    otros3?: number;
    serialp?: string;
    valido?: boolean;
    dis_cen?: string;
    aux01?: number;
    aux02?: string;
    ID?: number;
    salestax?: string;
    origen?: string;
    origen_d?: string;
    sta_prod?: string;
    telefono?: string;

    // relaciones
    reng_ped: RenglonPedidoDTO[];
}